Update Day 1
Today, I worked on writing code to figure out the number of occurrences for each category - placement, items, and team composition. My next goal is to cross-reference these occurrences. I want to see what my placement was for each occurrence of "trait" and "item". I'm not sure how to do it, but we will find out.