Update Day 2
Finished the function for getting specific placements, now trying to find a way to take the average and append that average to a new list in as few lines as possible. Going to do the same with the specific placement function. I think trait and traitlist might need to have the same syntax by means of modifying the data entry in excel. Im pretty sure I can condense at least the function down into only needing to be called once. I'm not sure how to initiate multiple lists in one line of code, but that may also be possible.