Update Day 2
Combined the functions for creating initial lists and the ones that both give a printable name and the average of the lists. Still trying to think of a way to further condense lines 95-128. I think there's a way to make one function that iterates through everything, but I can't think of it at the moment.